backend/requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
httpx
slowapi
pydantic-settings
pydantic[email]
alembic
opencv-python
mediapipe
ultralytics
jinja2
reportlab
numpy
pyyaml
python-multipart
openai>=1.12.0
croniter
pytesseract
